correct BlocksModule placement of HookHelper -> HookContainer. And refactor to new Core-2.0 Hook definition spec.
