+++ src/system/BlocksModule/composer.json
@@ -28,7 +28,7 @@
             "oldnames": [],
             "capabilities": {
                 "admin": {"route": "zikulablocksmodule_admin_view"},
-                "hook_subscriber": {"enabled": true}
+                "hook_subscriber": {"class": "Zikula\\BlocksModule\\Container\\HookContainer"}
             },
             "securityschema": {
                 "ZikulaBlocksModule::": "Block key:Block title:Block ID",
